feat: 챌린지, 모달 분리
